OscillatorB: Set xmod modulation offset to 0.0659
Set a modulation offset on the xmod signal, which results
in virtually zero pitch shift when the mod amount is
at 100 (max = 24 half semitones = 12 semitones), compared
to when it is 0. This is to increase the directness of the
xmod parameter, so that the oscillators don't need to be
completely retuned when advancing the control. There's still
a bit of detune that varies throughout the range.

Experiments were done bumping max xmod from current 24 half-
semitone steps to 100, but it turns out that above 24, we get
various forms of aliasing as the carrier oscillator gets into
the supersonic range, so it was deemed not worth while.