Implement key assignemnt modes from MiMi-a
The MiMi-a voice allocation is based on two priority queues,
onpri and offpri. The offpri queue contains all voices (pointers
to all voices) who are in the 'key off' state, and the onpri
queue contains all voices (pointers to all voices) who are in
the 'key on' state. There is also a stack of notes (unplayed_stack)
which contains notes that have not been played due to no free
voices being available according to the allocation rules currently
in effect, and an accompanying list of corresponding note-on
velocities for these notes.

In unison mode, the priority queues are not used as such, since
there is only a single note to keep track of, which is done using
a couple of simple variables, however, they are still used
as references to the Voice * array, providing a seamless transition
between unison and poly modes if the parameter is changed why
notes are being played.

When a note on event occurs, a voice fetch is attempted, according to
the following rules ('taking' a voice means removing (extracting)
it from the queue in which it was found):

1. If there is a free voice in offpri:
1a. If MEM (MEMory) is set, attempt to take a voice that has played
    the same note as the new note requested,
1a. otherwise, if RSZ (ReSet to Zero) is set, take the lowest
    numbered voice,
1c. otherwise, take the oldest = first voice in the queue (cyclic
    mode),
2. otherwise, if ROB O (Voice rob oldest) is set, take the first
   (oldest) voice in onpri. If ROB NL (Voice rob next-to-lowest)
   is set, instead take the next-to-lowest note in onpri.

If a voice is indeed taken, push it onto onpri, and gate it on.
If the voice was already playing, and Single Trig Note On is set,
do not retrigger the voice, but only change the pitch (legato).

If a voice was not found, push the note onto the unplayed note stack.
If a voice was robbed, push the note it was playing onto the unplayed
voice stack.

When a note off event occurs:

1. If the voice is found in onpri,
1a. and RES (Restore) is set and there is a note in the unplayed stack,
    pop the note off stack, push the voice onto onpri, and unless
    the Single Trig Note Off bit is set, retrigger the voice.
2b. otherwise, gate it off.
2. otherwise the note is not being played and no action is taken.

With these modes, everything from classic monophonic last note
priority to polyfonic cyclic with voice rob and memory modes can
be enabled.

Note that offpri and onpri do not take into account if the sustain
pedal was pressed, or if the voice has actually terminated
(reached the end of its useful life due to the loudness decay,
sustain or release phase reaching 0), but only if the corresponding
key is pressed or not.

In unison mode, use simplified allocation routines to avoid making
the poly routines too complex.

Params ASPLAYEDALLOCATION and LEGATO are now _NOTUSED

Remove VoiceQueue.h as it is not needed any more.

Remove now unused and previously useless anyway unisonOn function.