Voice: Always update LFOs even in economy mode
To keep the now-per-voice LFOs in phase with each other, they need
to be run at all times even when the voice is not playing.
The penalty is very small, as the update basically just is an
addition of the phase increment (plus one call to the random number
generator per cycle for the s/H waveform).