Lfo: Move sh generation to getValue()
Generate sample and hold waveform in getValue() like the other
waveforms. This makes the waveform generation consistent across
different waveforms, and also allows multiple LFOs to run off
a common phase if needed.

It also minimizes the amount of processing done in the update()
call, which is important, as this call is always performed even
when the voice is not playing, to maintain LFO phase sync.