Voice: Make envelope spread symmetric
Since the envelope spread parameter is applied multiplicatively,
as originally written, the range is 1 +/- 0.5 = 0.5 .. 1.5, i.e.
at the minimum value (0.5) it represents "half" whereas the maximum
value (1.5) represents a mere 50% increase and not a doubling.

This means that when applying spread, the average envelope times
will be slower than without the spread (it is the raw envelope values,
i.e. the phase increase, which are modified, not the actual A/D/ST/R
parameters, so lower values represent slower times).

Fix this by setting the range to 2 ** (rnd * spread), which will
result in a symmetric offset range when applied multiplicatively.