Voice: Add HPF, repurposing Brightness parameter
Add controllable high pass filter after LPF. Cutoff frequency
can be set from 4 Hz (which emulates a reasonably large
coupling cap after the filter, the most important effect of
which is to limit the amplitude of the self oscillating filter
at low cutoff frequencies) to 2500 Hz (which is slightly more
than practically useful - for instance, the HPF in the Juno-106
can be set to 0, 250 and 700 Hz).

Repurpose Brightness parameter as HPF cutoff freq.

TODO: Remove setBrightness and associated brightness variables.