Oscillators: Implement variable sync level
With variable sync level, the slave oscillator will only reset
if it happens to be above the sync level when the hard sync
occurs. This was implemented on the Delta Seven, but for various
reasons not carried through to the MiMi-a, so here we revive it!

Therefore, repurpose the boolean sync parameter as a variable
sync level from 0 (= off) to 1 (= full hard sync).

Note that the SyncLevel variable in Oscillators.h is actually
1 - param, so that 0 => full hard sync and 1 => no sync.