Oscillators et al: Add osc 2 detune
The MiMi-a has fine tune for both oscillators so the MiMi-d
should too.

In the UI, put the detune parameter where sync was, on the osc 2
page, and sync on the osc common page, so repurpose the osc sync
parameter number as osc 2 detune, and add a new parameter number
for ocs sync.