Remove general envelope generator
While it is sad to see it go, the General Envelope Generator has
been replaced by the new LFO/mod env design. Just as it came about
as a result of the original MiMi-a design having a general mod
output from the filter envelope, which because it needed its own
amount control turned into a fully fledged envelope, so does the
usefulness of a two LFO come mod env design mean that
the General Envelope Generator has become redundant.

It was also noticed on the MiMi-a how most of the time the GEG
was set up as a simple AD envelope, or used in repeat mode as
a second LFO. It was useful as a mod envelope / LFO because of
the MiMi-a linear envelope design, but the exponential envelopes
of the MiMi-d make this less attractive.

The genv controller source for the LFOs is now replaced with
velocity.

Since we now don't expect any fast-moving modulation amounts for
the LFOs, remove the modulation amount delay lines, just delay the
complete LFO signal.