Voice et al: Implement envelope RST mode
Envelope RST means that the envelopes will reset each time
the attack phase starts, thus every cycle of the envelope
will be identical, not depending on exactly where the
previous decay, sustain or release cycle left off.

Note that this is dependent on the key assign mode; if the
currently selected key assignment does not trigger the
envelope (legato modes) it will not cause a retrigger
even when envelope RST is set.