Voice: Ramp down delay line content when envelope reset
This minimizes clicks from envelopes which are reset from a fairly
high envelope level.

Instead of linear ramp, which itself gives a vast improvement
in click reduction when envelopes reset, attempt to make the ramp
even less audible by using a cosine function instead of a linear ramp.
The derivative of the cosine will be a sine pulse, rather than the
rectangle pulse which is the derivative of a linear ramp.

In order to avoid all delay lines having to calculate a ramp,
create separate DelayLineRampable class for use when this feature
is actually needed (amplifier and filter envelopes).

TODO: There is still sometimes a remaining click. Should the
DelayLineRampable class be used for other delay lines as well
besides the loudness and filter envelopes?