AdssrEnvelope: Remove time compensation
Remove the time compensation which attempts to have the
A, D, ST and R parameters set the time rather than the rate
of the respective envelope segment.

The main problem with having time compensation is that that is not
how an analog envelope works; indeed, time parameters are implemented
simply with resistors in series with the grounded timing cap, fed
from constant voltage sources, which set the rate of change, not the
time of change, whatever the front panel says. Also, since the curves
are exponential decay curves, the time of the curve seen in relative
terms (i.e. from a given percentage of the starting voltage to
the final percentage) is in fact constant, if not in terms of
absolute values.

Finally, the specific problem with the current time compensation
is that it fails to work properly, especially at the extremes
of the sustain level parameter, causing envelope control to appear
mildly haphazard.