AdssrEnvelope: Set decay asymptote to sustain level, not zero
The decay curve had zero as its asymptote, but should be the
sustain level, so fix that. Otherwise, the decay curve will have
an abrupt termination which gets worse the higher the sustain
level is (similar to the way the attack asymptote is higher
than the peak, but that is by design to a) get a well defined
trigger point in an analog envelope and b) to give a more
linear attack curve which is often desireble).

As a consequence, change the diff to sustain asymptote to change
0.01 which is small enough to get a good portion of the
exponential envelope shape while at the same time providing
a good trigger point.

The SUS state is now only used for the decaying sustain time
portion of the envelope, in ADSSR mode. In ADSR mode, it
remains in the DEC stage, making continued control of the
sustain level by the sustain control (with the decay time
providing lag).

As a bonus, the sustain level control now always controls the
sustain level via the decay time control, just like an analog
ADSR usually works (as long as the sustain phase has been reached).