AdssrEnvelope: In ADSSR mode, lower decay asymptote
Instead of adding an offset to the trigger point, which means that
at high sustain levels, the decay phase will be skipped and sustain
entered directly, lower the sustain asymptote by a small amount.
This decreases the decay time slightly because of the larger delta
between the attack level and trigger point, but gives more well
defined trigger point for the D->S transition, and also, perhaps
more importantly, makes the decay phase perceptually correspond
to the part of the decay curve which is most prominent, i.e.
before the exponential "tail".

In order to compensate for the lower sustain asymptote in ADDSR mode,
lower the decay coefficient thus lengthening the envelope segment, to
compensate for the fact that it will decay faster since the asymptote
is lower. The calculations are based on the values at the start
of the decay phase, when the output value is 1.

TODO: Need to remove unused2 and make asymptote delta onstant.