AdssrEnvelope: Add linear envelope mode
This mimics the MiMi-a envelopes which are always linear.

In the MiMi-d case, try to match the linear envelope times to
their exponential counterparts, especially the attack and decay phases.

Note that the sustain asymptote in linear mode needs to be
the same as the sustain level, in contrast to exponential mode
where it is slightly lower.