ParamDefs: Move default values (+ min, max) to ParamDefs
Move default values from Params.h to ParamDefs (with some minor
adjustments), and include it in Params class in such a way that
the default values are set by the constructor.

Min and max values are also set (all parameters are set to
min = 0.0 and max = 1.0), but are currently not used.